% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIDconverter2.R
\name{createIDconverter2}
\alias{createIDconverter2}
\title{ID Converter Function}
\usage{
createIDconverter2(annotationPackageName, from, to)
}
\arguments{
\item{annotationPackageName}{Annotation package name. Format = String.}

\item{from}{Input ID type. See keytypes(<annotation package name>) for annotation package-specific ID types}

\item{to}{Output ID type. See keytypes(<annotation package name>) for annotation package-specific ID types}
}
\description{
This function takes an annotation package, and conbversion types as input to create an ID converter function
This code was modified from the SetRank package (Cedric Simillion (2016). SetRank: Advanced Gene Set Enrichment Analysis. R package version 1.1.0. https://CRAN.R-project.org/package=SetRank)
}
\author{
Brandon D. Murugan
}
