% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{data_preprocess}
\alias{data_preprocess}
\title{Imports and preprocesses data files for label-free proteomic LC-MS experiments}
\usage{
data_preprocess(txtpath, mqparfile, quant_type = "Intensity",
  norm_type = "MedianCentre", valid_values = 3, imputation = F)
}
\arguments{
\item{txtpath}{MaxQuant txt folder}

\item{quant_type}{Quanitiation type to select. Possible choices are "Intensity" (default), or "LFQ".}

\item{norm_type}{Type of normalisation to perform. Possible choices are "MedianCentre" (default), "Quantile", or "None".}

\item{valid_values}{Numeric. Removes entire protein if number of valid (non-zero) values are less than stated amount for \bold{any} group}

\item{imputation}{Logical. If \code{TRUE}, will impute with lowest value of all intensities found in dataset. Default is FALSE}

\item{mqparPath}{MaxQuant mqpar.xml file path}
}
\description{
This function imports MaxQuant proteingroups.txt, peptides.txt, Phospho (STY)Sites.txt (if it exists), and mqpar.xml files.
Proteingroups.txt file is then filtered and normalised according to chosen quantitation and normlisation types
}
\author{
Brandon D. Murugan
}
