% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnhancedGraphicsReformat.R
\name{EnhancedGraphicsReformat}
\alias{EnhancedGraphicsReformat}
\title{Enhanced Graphics to add modifications to WikiPathways in Cy3 ####}
\usage{
EnhancedGraphicsReformat(input.df)
}
\arguments{
\item{input.df}{Input dataframe. Must have columns as follows:
\bold{ID:} JUN-S73
\bold{Symbols:} JUN
\bold{Sites:} S73
\bold{Effect:} NA
\bold{log2FC:} 0.5555}
}
\description{
This function takes a file with ___ columns, and reformats it for import into Cytoscape for use with the \bold{EnhancedGraphic} package.
}
\author{
Brandon D. Murugan
}
