% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_valid_values.R
\name{filter_valid_values}
\alias{filter_valid_values}
\title{Valid value filter for label-free proteomic LC-MS experiments}
\usage{
filter_valid_values(
  input.df,
  experimental.groups,
  valid.values = 2,
  ID.columns = 1,
  zero.as.na = T
)
}
\arguments{
\item{input.df}{Input dataframe. Must have one column with protein/gene identifier and rest of columns with expression values for all samples}

\item{experimental.groups}{List of experimental groups with replicate suffix excluded (e.g.: c("Control", "Treatment"), \bold{not} c("Control_1", "Control_2", etc). This list must match at least part of the intensity column names in the expression matrix)}

\item{valid.values}{Number of valid values to filter for per-group}

\item{ID.columns}{Number of ID columns. Will assume the 1st n columns (Default = 1)}

\item{zero.as.na}{Returns dataframe with zeros converted to NAs (Default = T)}
}
\description{
This function takes a file with protein names 1st column) and quantitation columns, and applies a user defined valid value filter with per-group constraints.
The filter will assess each group for the minimum number of valid values, and return the rows where this filter passes in \bold{all} groups
The output file is the expression matrix with rows that do \bold{not} pass the filter removed
}
\author{
Brandon D. Murugan
}
