% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.name.extractor.R
\name{gene.name.extractor}
\alias{gene.name.extractor}
\title{Gene Name Extractor}
\usage{
gene.name.extractor(inputDF, SwissProt.only = T)
}
\arguments{
\item{inputDF}{Input MQ proteomic dataframe with Fasta.headers column}

\item{SwissProt.only}{Will select first SwissProt entry in ProteinGroups. (Default = TRUE)}
}
\description{
This function Extracts the Gene Name and Uniprot Identifiers from the Fasta.header column of the ProteinGroups file.
Will also extract the appropriate Uniprot ID.
Can select first SwissProt entry in protein group if required
}
\author{
Brandon D. Murugan
}
