% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowest_taxonomic_level.R
\name{lowest.tax.level}
\alias{lowest.tax.level}
\title{Lowest Taxonomic Level for each OTU}
\usage{
lowest.tax.level(phyloseq.file)
}
\arguments{
\item{phyloseq}{Input phyloseq object}
}
\description{
This function takes in a phyloseq object and adds a column in the tax_table with the lowest taxonomic level for each OTU/ASV.
}
\details{
Returns new column in tax_table names \bold{lowest_tax_level}
}
\author{
Brandon D. Murugan
}
