% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediancentre.R
\name{mediancentre}
\alias{mediancentre}
\title{Mediancentre function}
\usage{
mediancentre(x)
}
\description{
This function reads in a dataframe with protein ID column as the 1st column
}
\details{
median centre normalisation:
data distributions are adjusted such that each individual sample's medians are aligned to the average of medians of all the samples
}
