% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBIMappingData.R
\docType{data}
\name{NCBI_200210}
\alias{NCBI_200210}
\title{NCBI Human mapping file}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://www.ncbi.nlm.nih.gov/gene}{NCBI}
}
\usage{
data(NCBI_200210)
}
\description{
NCBI data for human genome (GRCh38.p13)
Genome Reference Consortium (March 2019)
Data extracted on 10 February 2020
}
\examples{
None
}
\keyword{NCBI}
