% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBIMappingData_Assembly.R
\docType{data}
\name{NCBI_assembly_200313}
\alias{NCBI_assembly_200313}
\title{NCBI Human mapping file - Assembly}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://www.ncbi.nlm.nih.gov/gene}{NCBI}
}
\usage{
data(NCBI_assembly_200313)
}
\description{
RefSeq data for human genome (GRCh38.p13)
Genome Reference Consortium (March 2019)
Data extracted on 13 March 2020
}
\examples{
None
}
\keyword{NCBI}
