% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBIMappingData_Genes.R
\docType{data}
\name{NCBI_genes_200210}
\alias{NCBI_genes_200210}
\title{NCBI Human mapping file - Genes}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://www.ncbi.nlm.nih.gov/gene}{NCBI}
}
\usage{
data(NCBI_genes_200210)
}
\description{
NCBI Genes
Data extracted on 10 Feb 2020
}
\examples{
None
}
\keyword{NCBI}
