% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBIMappingData_Genes_unique.R
\docType{data}
\name{NCBI_genes_200210_unique}
\alias{NCBI_genes_200210_unique}
\title{NCBI Human mapping file - Genes: Filtered}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/gene}{NCBI}
}
\usage{
data(NCBI_genes_200210_unique)
}
\description{
NCBI Genes
Data extracted on 10 Feb 2020
Duplicates removed
}
\examples{
None
}
\keyword{NCBI}
