% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.select.R
\name{sp.select}
\alias{sp.select}
\title{Protein ID Filter}
\usage{
sp.select(
  input.df,
  ProteinID.column,
  new.ID.column = "ProteinID_main",
  delimiter
)
}
\arguments{
\item{ProteinID.column}{Protein ID column to be filtered through to find 1st instance of an \bold{sp} identifier (string)}

\item{new.ID.column}{Name of new ID column. Default is \bold{ProteinID_main} (string)}

\item{delimiter}{Delimiter within \bold{ProteinID.column} (string)}
}
\description{
This function takes a file with protein IDs and returns the 1st \bold{sp} (SwissProt) identifier. The function will return the 1st \bold{tr} (trembl) entry if no sp entries are found.
}
\details{
The output file is the input file with an additional \bold{user defined} ProteinID column (Default = \bold{ProteinID_main})
}
\author{
Brandon D. Murugan
}
