% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotMappingData.R
\docType{data}
\name{Uniprot2Genename_swissprot_20200224}
\alias{Uniprot2Genename_swissprot_20200224}
\title{UniProt Human GeneName to UniProtKB mapping file}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://www.uniprot.org}{UniProtKB}
}
\usage{
data(Uniprot2Genename_swissprot_20200224)
}
\description{
UniProt Human Reference proteome filtered for SwissProt entries only (Reviewed)
Data extracted on 24 Feb 2020
}
\examples{
None
}
\keyword{uniprot}
