% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.select.R
\name{sp.select}
\alias{sp.select}
\title{Protein ID Filter}
\usage{
sp.select(input.df, ProteinID.column, new.ID.column = "ProteinID_main",
  delimiter)
}
\arguments{
\item{input.df}{Input dataframe (e.g.: MaxQuant proteinGroups.txt file) \italic{(environmental variable)}}

\item{ProteinID.column}{Protein ID column to be filtered through to find 1st instance of an \bold{sp} identifier \italic{(string)}}

\item{new.ID.column}{Name of new ID column. Default is \bold{ProteinID_main} \italic{(string)}}

\item{delimiter}{Delimiter within \bold{ProteinID.column} \italic{(string)}}
}
\description{
This function takes a file with protein IDs and returns the 1st \bold{sp} (SwissProt) identifier. The function will return the 1st \bold{tr} (trembl) entry if no sp entries are found.
}
\details{
The output file is the input file with an additional \italic{user defined} ProteinID column (Default = \bold{ProteinID_main})
}
\author{
Brandon D. Murugan
}
